refactor: 작성한 페이지 리팩토링
- 각 페이지 AppHelmet 적용
- MyPage 프로필 이미지 대체 텍스트 변경
- 페이지 컴포넌트 폴더 구조 정리
