+++ src/pages/MyPage/index.jsx
@@ -1,14 +1,14 @@
 import getPbImageURL from '@/api/getPbImageURL';
 import pb from '@/api/pb';
+import AppHelmet from '@/components/AppHelmet/AppHelmet';
 import AppSpinner from '@/components/AppSpinner/AppSpinner';
 import BasicTextModal from '@/components/BasicTextModal/BasicTextModal';
 import CommonBtn from '@/components/CommonBtn/CommonBtn';
 import useGlobalStore from '@/stores/useGlobalStore';
 import useModalStore from '@/stores/useModalStore';
 import { useEffect, useState } from 'react';
-import { Helmet } from 'react-helmet-async';
 import { Link, useNavigate } from 'react-router-dom';
-import ProfileBox from './components/ProfileBox';
+import ProfileBox from './components/ProfileBox/ProfileBox';
 import S from './MyPage.module.css';
 
 export function Component() {
@@ -88,20 +88,7 @@ export function Component() {
 
   return (
     <>
-      <Helmet>
-        <title>마이트립픽 | 마이페이지</title>
-        <meta
-          name="description"
-          content="나만의 인생 여행지를 올려 다른 사람들과 공유하고, 가장 높은 순위를 가진 여행지를 선정해 보세요."
-        />
-        <meta property="og:title" content="마이트립픽" />
-        <meta property="og:type" content="site" />
-        <meta
-          property="og:description"
-          content="나만의 인생 여행지를 올려 다른 사람들과 공유하고, 가장 높은 순위를 가진 여행지를 선정해 보세요."
-        />
-        <meta property="og:site:author" content="리액트에서-구해조" />
-      </Helmet>
+      <AppHelmet title="마이트립픽 | 마이페이지" />
       <h1 className="a11yHidden">마이페이지</h1>
       <ProfileBox userData={profileData} />
       <section>
