+++ src/pages/PostDetail/components/PostText/PostText.jsx
